+++ app/Basket/Synchronisation/ApplicationSynchronisationService.php
@@ -385,4 +385,27 @@ class ApplicationSynchronisationService extends AbstractSynchronisationService
             throw $e;
         }
     }
+
+    /**
+     * @author SL
+     * @param Application $application
+     * @return array
+     * @throws \Exception
+     */
+    public function getApplicationHistory(Application $application)
+    {
+        try {
+            return $this->applicationGateway->getApplicationHistory(
+                $application->ext_id,
+                $application->installation->merchant->token
+            );
+        } catch (\Exception $e) {
+
+            $this->logError(
+                'ApplicationSynchronisationService: Fetching Application Status History failed: ' . $e->getMessage()
+            );
+
+            throw $e;
+        }
+    }
 }